fix more shape list
